Rename ``VQEProgram`` to ``VQEClient`` (#351)
* rename VQEProgram to VQERuntimeClient

* black

* add reno

* comments from code review (round 1)

* extend tests to check result type of fakeruntime

* comments from code review

- use deprecated module
- rename  to VQEClient
- use combine to combine result objects

* fix typos nature<->optimization

* update the tutorial

* update tutorial

* add "optimizers" to pylintdict

* Deprecate for 0.3.0


* rename file to vqe_client

* follow up on merge

* use "minimum version numbers" instead of fixed ones
