+++ .pylintdict
@@ -280,6 +280,7 @@ oneeints
 onwards
 opflow
 oppenheimer
+optimizers
 orbsym
 ortho
 orthonormal
